feat: add location
it works, but we need to regenerate the template bundle with

`go generate ./cmd/fyne/internal/templates/...`

we also need to regenerate with gendex.
for that, we need a version of Android SDK that still has the `dx` utility.

currently it works with:
build-tools/30.0.3
platforms/android-30

and

```
go run -tags gendex ./cmd/fyne/internal/mobile/gendex
go install ./cmd/fyne
```